refactor(substrate): add a test-support feature hosting the canonical TestChassis fixture
Consolidate the per-crate `TestChassis` fixture copies into one
`#[cfg(any(test, feature = "test-support"))] pub mod testing` in
aether-substrate, reached by downstream crates through a `test-support`
dev-dependency feature.

The module is the capabilities superset moved verbatim (`TestChassis`,
`fresh_substrate`, `test_mailer_and_rx`, `drive_task_completion`, the
reply decoders, `scratch_dir` / `cleanup`) plus three items that let
every prior copy collapse without changing what any test exercises:
`bare_substrate` carries `test_util::fresh_substrate`'s narrow semantics
(no descriptors, no outbound), `fresh_substrate_and_rx` exposes the egress
receiver (`fresh_substrate` wraps it, dropping the rx), and
`boot_test_chassis_with` folds in from labyrinth.

- aether-substrate: new `test-support` feature + `src/testing.rs`;
  `test_util.rs` retired; the internal fixture sites (including the inline
  `TestChassis` in `chassis/builder.rs` and `tests/native_actor_macro.rs`)
  repointed to `crate::testing` / `aether_substrate::testing`.
- aether-capabilities / aether-mcp / aether-labyrinth: local `test_chassis`
  modules deleted, imports repointed, dev-dep gains the feature. Labyrinth's
  local seed folds into `fresh_substrate_and_rx` (tuple order canonicalized).
- aether-substrate-bundle: three inline integration-test impls replaced by
  the canonical import; dev-dep gains the feature.
