+++ crates/aether-substrate/src/testing.rs
@@ -1,15 +1,25 @@
-//! Shared `TestChassis` fixture for unit tests across `aether-capabilities`.
+//! Canonical cross-crate test fixture for booting a chassis in tests.
 //!
-//! Every cap's `#[cfg(test)] mod tests` exercises its `init` / handlers
-//! by booting a real [`Builder`] against a no-op chassis declaration. Pre-extraction every site copied
-//! the same 8-line `impl Chassis for TestChassis` block; this module is
-//! the single canonical declaration so test modules just
-//! `use crate::test_chassis::TestChassis;` and address it by the typename
-//! the builder expects.
+//! Every crate that exercises a native cap's `init` / handlers boots a
+//! real [`Builder`] against a no-op [`TestChassis`] declaration. This
+//! module is the single home for that fixture plus the seed / decode /
+//! drive helpers, so downstream crates reach it through the
+//! `test-support` dev-dependency feature instead of re-declaring their
+//! own copy. Substrate's own `#[cfg(test)]` modules reach it through the
+//! `test` arm of the module's cfg gate.
 //!
-//! Filed by issue 785. The `fresh_substrate` helper extracted by issue
-//! 786 lives here too — same six sites all wanted the same
-//! `(Arc<Registry>, Arc<Mailer>)` seed for `Builder::new`.
+//! Two seeds live here because substrate-internal tests and cap tests
+//! want different starting states: [`bare_substrate`] registers no kind
+//! descriptors and wires no outbound (substrate's own tests exercise
+//! neither descriptor lookup nor the ADR-0037 bubble-up path), while
+//! [`fresh_substrate`] / [`fresh_substrate_and_rx`] pre-populate the kind
+//! descriptors and wire a loopback outbound for the cap tests that do.
+
+#![allow(
+    clippy::must_use_candidate,
+    clippy::missing_panics_doc,
+    reason = "test-support fixtures: the returned seeds are always consumed by the caller, and the `expect`s on setup are themselves the test assertions"
+)]
 
 use std::env::temp_dir;
 use std::fs;
@@ -21,22 +31,22 @@ use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use aether_data::{Kind, Source};
 use aether_kinds::descriptors;
-use aether_substrate::actor::native::binding::NativeBinding;
-use aether_substrate::actor::native::ctx::NativeCtx;
-use aether_substrate::actor::native::{NativeActor, TaskCompletionWake};
-use aether_substrate::chassis::Chassis;
-use aether_substrate::chassis::builder::{BuiltChassis, NeverDriver};
-use aether_substrate::chassis::error::BootError;
-use aether_substrate::mail::mailer::Mailer;
-use aether_substrate::mail::outbound::{EgressEvent, HubOutbound};
-use aether_substrate::mail::registry::Registry;
+
+use crate::actor::native::binding::NativeBinding;
+use crate::actor::native::ctx::NativeCtx;
+use crate::actor::native::{NativeActor, TaskCompletionWake};
+use crate::chassis::Chassis;
+use crate::chassis::builder::{Builder, BuiltChassis, NeverDriver, PassiveChassis};
+use crate::chassis::error::BootError;
+use crate::mail::mailer::Mailer;
+use crate::mail::outbound::{EgressEvent, HubOutbound};
+use crate::mail::registry::Registry;
 
 /// Canonical test chassis. `build()` is unreachable — every consumer
 /// drives the chassis through `Builder::<TestChassis>::new(...)` directly
 /// rather than going through `TestChassis::build(())`.
 pub struct TestChassis;
 
-//noinspection DuplicatedCode
 impl Chassis for TestChassis {
     const PROFILE: &'static str = "test";
     type Driver = NeverDriver;
@@ -46,22 +56,59 @@ impl Chassis for TestChassis {
     }
 }
 
-/// Build the `(Arc<Registry>, Arc<Mailer>)` seed every cap test feeds to
-/// `Builder::<TestChassis>::new`. The registry is pre-populated with the
-/// substrate kind descriptors so tests can address built-in kinds by id
-/// without re-registering; the mailer carries a loopback `HubOutbound`
-/// (rx dropped) so the unknown-mailbox bubble-up path (ADR-0037) is
-/// wired but inert — tests that never hit it (audio, fs, http handler
-/// paths) see no behavioural difference, and tests that do hit it
-/// (rpc, engine proxy) get the connected backend they need.
-pub fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
+/// Build the `(Arc<Registry>, Arc<Mailer>)` seed substrate-internal tests
+/// feed to `Builder::<...>::new` and `NativeBinding::new_for_test`.
+/// Intentionally narrower than [`fresh_substrate`]: no kind descriptors
+/// registered, no outbound wired — substrate-internal tests don't exercise
+/// descriptor lookup or the unknown-mailbox bubble-up path (ADR-0037).
+pub fn bare_substrate() -> (Arc<Registry>, Arc<Mailer>) {
+    let registry = Arc::new(Registry::new());
+    let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
+    (registry, mailer)
+}
+
+/// Build the seeded `(Arc<Registry>, Arc<Mailer>, Receiver<EgressEvent>)`
+/// triple every cap test feeds to `Builder::<TestChassis>::new`, exposing
+/// the egress receiver so callers can drain `SourceAddr::Session` sends.
+/// The registry is pre-populated with the substrate kind descriptors so
+/// tests can address built-in kinds by id without re-registering; the
+/// mailer carries a loopback `HubOutbound` so the unknown-mailbox
+/// bubble-up path (ADR-0037) is wired but inert — tests that never hit it
+/// (audio, fs, http handler paths) see no behavioural difference, and
+/// tests that do hit it (rpc, engine proxy) get the connected backend
+/// they need.
+pub fn fresh_substrate_and_rx() -> (Arc<Registry>, Arc<Mailer>, Receiver<EgressEvent>) {
     let registry = Arc::new(Registry::new());
     for d in descriptors::all() {
         let _ = registry.register_kind_with_descriptor(d);
     }
-    let (outbound, _rx) = HubOutbound::attached_loopback();
+    let (outbound, rx) = HubOutbound::attached_loopback();
     let mailer = Arc::new(Mailer::new(Arc::clone(&registry)).with_outbound(outbound));
+    (registry, mailer, rx)
+}
+
+/// The seeded `(Arc<Registry>, Arc<Mailer>)` seed for cap tests that don't
+/// observe egress. A thin wrapper over [`fresh_substrate_and_rx`] that
+/// drops the receiver.
+pub fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
+    let (registry, mailer, _rx) = fresh_substrate_and_rx();
     (registry, mailer)
+}
+
+/// Boot a `TestChassis` carrying exactly one cap `A` with `config`. The
+/// minimal-boot path a cap's dispatcher-thread test reaches for.
+pub fn boot_test_chassis_with<A>(
+    registry: &Arc<Registry>,
+    mailer: &Arc<Mailer>,
+    config: A::Config,
+) -> PassiveChassis<TestChassis>
+where
+    A: NativeActor,
+{
+    Builder::<TestChassis>::new(Arc::clone(registry), Arc::clone(mailer))
+        .with_actor::<A>(config)
+        .build_passive()
+        .expect("test chassis boots")
 }
 
 /// Build a `(Arc<Mailer>, Receiver<EgressEvent>)` pair where the
@@ -162,9 +209,6 @@ where
 /// following egress is a `ToSession` carrying `K`, rather than reading
 /// past bubble-ups. For cap tests that drive via the full dispatcher and
 /// egress channel (e.g. `render-runtime`) rather than direct `-> R` calls.
-// Used by render.rs tests (feature = "render-runtime"); dead_code fires in
-// the default build without that feature.
-#[allow(dead_code)]
 pub fn decode_reply<K>(rx: &Receiver<EgressEvent>) -> K
 where
     K: Kind,
